create prototype
The prototype allows for user interactions via a popup that can be
triggered from a shortcut and accepts various key combinations.

So far you can pin tabs, move the focus to pinned tabs, and bring pinned
tabs to where the focus is. Each pin is associated with a key on the
keyboard.