further refactor the service worker
This commit introduces "views" of the keyset cache that can layer
keysets and provide operations on them. These views implement consistent
semantics of CRUDL operations with the goal of providing a user
experience with fewer surprises.

As a counterpart, this commit also introduces "key references", which
describe a key in a keyset. Key references are used to unambiguously
define keys in keysets for operations that cannot work with the layer
semantics described above. For that matter, the layer views use those
key references under the hood.