add a command combo to move pins
Since a move requires a source and a destination, this changes required
an update to the ComboTrie, so that it can parse multiple keyrefs.

This commit also fixes an existing move issue that would override keyset
0 as the destination keyset.

There is a known issue with this command, though: If you attempt to move
a pin across keysets without explicitly specifing the source keyset
(e.g., `ma2b` instead of `m1a2b`), then the code will use the
destination keyset as the target (i.e., `m2a2b`). I accept this flaw
given that it's easy to work around by specifying the source keyset
explicitly.