refactor ComboTrie
The `ComboTrie` has been populated only in the `popup.js`. It should
also be available for shortcuts: A shortcut can be programmed with a
combo to execute when that shortcut is pressed. However, shortcuts are
handled in the service worker which has to deal with a slightly
different API and doesn't need to manage the popup window. To enable the
use of the `ComboTrie` in different contexts, this refactoring has been
done.