add the edit modal in the popup window
This side panel is unwieldy, requiring extra navigation and clicking to
get to the edit dialog of a pin. To this end, it's better to have the
modal directly in the popup.

This required an updated commmand to invoke the dialog as well as a
somewhat hacky workaround to not interpret keystrokes as popup commands
when the dialog is showing. Maybe there's a more elegant way to prevent
key events from bubbling above the modal but I save that for later.