fix side-effect issue with `listEntries()`
`listEntries()` was supposed to have a side effect (updating the pin),
but it actually created a new object. So the side effect would only
become effective on the next `listEntries()` invocation. For example,
this made it so that pins weren't correctly detected in the pop-up in
certain cases. Turning the (ostensible) side effect into a proper return
value solves this issue and improves readability of the code.