indicate currently active pin as badge text`
As the user navigates their tabs, a badge text at the action icon
indicates the currently selected tab's pin if any.

This change turned out to be more profound: Displaying the current pin
means, we need to continuously keep track of tabs and their pins -- not
only when the popup opens. As a result, this change introduced event
listeners for both changes to tabs and changes to pins.